requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
pymongo==4.6.1
python-dotenv==1.0.0
motor==3.3.2
aiohttp==3.8.5